+++ compiler-plugin/src/main/java/io/ballerina/stdlib/persist/compiler/utils/Utils.java
@@ -18,7 +18,12 @@
 
 package io.ballerina.stdlib.persist.compiler.utils;
 
+import io.ballerina.compiler.syntax.tree.AnnotationNode;
+import io.ballerina.compiler.syntax.tree.ExpressionNode;
+import io.ballerina.compiler.syntax.tree.MappingConstructorExpressionNode;
+import io.ballerina.compiler.syntax.tree.MappingFieldNode;
 import io.ballerina.compiler.syntax.tree.Node;
+import io.ballerina.compiler.syntax.tree.SpecificFieldNode;
 import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
 import io.ballerina.projects.plugins.codeaction.CodeActionArgument;
 import io.ballerina.projects.plugins.codeaction.CodeActionContext;
@@ -31,6 +36,7 @@ import io.ballerina.toml.syntax.tree.DocumentNode;
 import io.ballerina.toml.syntax.tree.KeyValueNode;
 import io.ballerina.toml.syntax.tree.NodeList;
 import io.ballerina.toml.syntax.tree.SyntaxTree;
+import io.ballerina.toml.syntax.tree.TableArrayNode;
 import io.ballerina.toml.syntax.tree.TableNode;
 import io.ballerina.tools.diagnostics.Diagnostic;
 import io.ballerina.tools.diagnostics.DiagnosticProperty;
@@ -44,9 +50,13 @@ import org.wso2.ballerinalang.compiler.diagnostic.properties.BStringProperty;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
+import java.util.Optional;
+import java.util.stream.Stream;
 
 /**
  * Class containing util functions.
@@ -175,25 +185,40 @@ public final class Utils {
         if (balProjectDir == null) {
             throw new BalException("unable to locate the project's Ballerina.toml file");
         }
-
         Path configPath = balProjectDir.resolve(ProjectConstants.BALLERINA_TOML);
+        Path targetDir = Paths.get(String.valueOf(balProjectDir), "target");
+        Path genCmdConfigPath = targetDir.resolve("Persist.toml");
+        if (Files.exists(genCmdConfigPath)) {
+            configPath = genCmdConfigPath;
+        }
+        return getDataStoreName(configPath);
+    }
+
+    private static String getDataStoreName(Path configPath) throws BalException {
         try {
             TextDocument configDocument = TextDocuments.from(Files.readString(configPath));
             SyntaxTree syntaxTree = SyntaxTree.from(configDocument);
             DocumentNode rootNote = syntaxTree.rootNode();
             NodeList<DocumentMemberDeclarationNode> nodeList = rootNote.members();
             for (DocumentMemberDeclarationNode member : nodeList) {
-                if (member instanceof TableNode) {
-                    TableNode node = (TableNode) member;
-                    String tableName = node.identifier().toSourceCode().trim();
+                if (member instanceof TableArrayNode arrNode) {
+                    String tableName = arrNode.identifier().toSourceCode().trim();
+                    if (tableName.equals(Constants.TOOL_PERSIST)) {
+                        for (KeyValueNode field : arrNode.fields()) {
+                            if (field.identifier().toSourceCode().trim().equals(Constants.OPTIONS_DATASTORE)) {
+                                return field.value().toSourceCode().trim().replaceAll("\"", "");
+                            }
+                        }
+                    }
+                } else if (member instanceof TableNode tableNode) {
+                    String tableName = tableNode.identifier().toSourceCode().trim();
                     if (tableName.equals(Constants.PERSIST)) {
-                        for (KeyValueNode field : node.fields()) {
+                        for (KeyValueNode field : tableNode.fields()) {
                             if (field.identifier().toSourceCode().trim().equals(Constants.DATASTORE)) {
                                 return field.value().toSourceCode().trim().replaceAll("\"", "");
                             }
                         }
                     }
-
                 }
             }
             throw new BalException("the persist.datastore configuration does not exist in the Ballerina.toml file");
@@ -216,28 +241,32 @@ public final class Utils {
         }
 
         Path configPath = balProjectDir.resolve(ProjectConstants.BALLERINA_TOML);
-        try {
-            TextDocument configDocument = TextDocuments.from(Files.readString(configPath));
-            SyntaxTree syntaxTree = SyntaxTree.from(configDocument);
-            DocumentNode rootNote = syntaxTree.rootNode();
-            NodeList<DocumentMemberDeclarationNode> nodeList = rootNote.members();
-            for (DocumentMemberDeclarationNode member : nodeList) {
-                if (member instanceof TableNode) {
-                    TableNode node = (TableNode) member;
-                    String tableName = node.identifier().toSourceCode().trim();
-                    if (tableName.equals(Constants.PERSIST)) {
-                        for (KeyValueNode field : node.fields()) {
-                            if (field.identifier().toSourceCode().trim().equals(Constants.DATASTORE)) {
-                                return field.value().toSourceCode().trim().replaceAll("\"", "");
-                            }
+        return getDataStoreName(configPath);
+    }
+
+    public static List<String> readStringArrayValueFromAnnotation(List<AnnotationNode> annotationNodes,
+                                                                  String annotation, String field) {
+        for (AnnotationNode annotationNode : annotationNodes) {
+            String annotationName = annotationNode.annotReference().toSourceCode().trim();
+            if (annotationName.equals(annotation)) {
+                Optional<MappingConstructorExpressionNode> annotationFieldNode = annotationNode.annotValue();
+                if (annotationFieldNode.isPresent()) {
+                    for (MappingFieldNode mappingFieldNode : annotationFieldNode.get().fields()) {
+                        SpecificFieldNode specificFieldNode = (SpecificFieldNode) mappingFieldNode;
+                        String fieldName = specificFieldNode.fieldName().toSourceCode().trim();
+                        if (!fieldName.equals(field)) {
+                            return Collections.emptyList();
+                        }
+                        Optional<ExpressionNode> valueExpr = specificFieldNode.valueExpr();
+                        if (valueExpr.isPresent()) {
+                            return Stream.of(valueExpr.get().toSourceCode().trim().replace("\"", "")
+                                    .replace("[", "")
+                                    .replace("]", "").split(",")).map(String::trim).toList();
                         }
                     }
-
                 }
             }
-            throw new BalException("the persist.datastore configuration does not exist in the Ballerina.toml file");
-        } catch (IOException e) {
-            throw new BalException("error while reading persist configurations. " + e.getMessage());
         }
+        return Collections.emptyList();
     }
 }